requirements.txt
absl-py==0.13.0
appnope==0.1.2
astunparse==1.6.3
attrs==21.2.0
backcall==0.2.0
blis==0.7.4
cachetools==4.2.2
catalogue==2.0.6
certifi==2021.5.30
charset-normalizer==2.0.4
clang==5.0
click==7.1.2
cupy-cuda112==9.4.0
cymem==2.0.5
datasets==1.11.0
debugpy==1.4.3
decorator==5.1.0
dill==0.3.4
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.1.0/en_core_web_sm-3.1.0-py3-none-any.whl
entrypoints==0.3
es-core-news-lg @ https://github.com/explosion/spacy-models/releases/download/es_core_news_lg-3.1.0/es_core_news_lg-3.1.0-py3-none-any.whl
es-core-news-md @ https://github.com/explosion/spacy-models/releases/download/es_core_news_md-3.1.0/es_core_news_md-3.1.0-py3-none-any.whl
es-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/es_core_news_sm-3.1.0/es_core_news_sm-3.1.0-py3-none-any.whl
es-dep-news-trf @ https://github.com/explosion/spacy-models/releases/download/es_dep_news_trf-3.1.0/es_dep_news_trf-3.1.0-py3-none-any.whl
fastrlock==0.6
filelock==3.0.12
flatbuffers==1.12
fsspec==2021.8.1
gast==0.4.0
google-api-core==2.0.1
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.0.0
google-cloud-translate==3.4.0
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grpcio==1.40.0
h5py==3.1.0
huggingface-hub==0.0.16
idna==3.2
ipykernel==6.4.1
ipython==7.27.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==7.0.2
jupyter-core==4.7.1
kaleido==0.2.1
keras==2.6.0
Keras-Preprocessing==1.1.2
language-tool-python==2.6.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
multiprocess==0.70.12.2
murmurhash==1.0.5
nbformat==5.1.3
nest-asyncio==1.5.1
nltk==3.3
numpy==1.21.2
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.2
parso==0.8.2
pathy==0.6.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
plotly==5.3.1
preshed==3.0.5
prompt-toolkit==3.0.20
proto-plus==1.19.0
protobuf==3.17.3
ptyprocess==0.7.0
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycountry==20.7.3
pydantic==1.8.2
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.2.1
regex==2021.8.28
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
sacremoses==0.0.45
scikit-learn==0.24.2
scipy==1.7.1
sentencepiece==0.1.96
six==1.16.0
sklearn==0.0
smart-open==5.2.1
spacy==3.1.2
spacy-alignments==0.8.3
spacy-legacy==3.0.8
spacy-transformers==1.0.6
spacy-wordnet @ git+https://github.com/recognai/spacy-wordnet.git@ada107749d0e195dd966a2d4afe94136cdb230ab
srsly==2.4.1
tenacity==8.0.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
thinc==8.0.10
threadpoolctl==2.2.0
tokenizers==0.10.3
torch==1.9.0+cu111
torchaudio==0.9.0
torchvision==0.10.0+cu111
tornado==6.1
tqdm==4.62.2
traitlets==5.1.0
transformers==4.9.0
typer==0.3.2
typing-extensions==3.10.0.2
urllib3==1.26.6
wasabi==0.8.2
wcwidth==0.2.5
Werkzeug==2.0.1
wrapt==1.12.1
xxhash==2.0.2
